Move App.tsx to the root folder and remove unused
