+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React, { createContext } from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
-import App from "./components/App/App";
+import App from "./App";
 import Store from "./store/store";
 import { BrowserRouter as Router } from "react-router-dom";
